Leaf option for blocks
Adds the option to set a block attribute as a leaf. Leaf blocks cannot contain other nested blocks.

Useful for stuff like headings. When the current block is an H1 and i press H2 I would expect the H1 to be replaced by an H2, not nested in it.
Fixes https://github.com/basecamp/trix/issues/133#issuecomment-166155477